move `generate_privacy_manifest!`
