+++ ios/pod_helpers.rb
@@ -1,3 +1,15 @@
+require('json')
+
+def app_manifest(project_root)
+  @app_manifest ||= {}
+  return @app_manifest[project_root] if @app_manifest.key?(project_root)
+
+  manifest_path = find_file('app.json', project_root)
+  return if manifest_path.nil?
+
+  @app_manifest[project_root] = JSON.parse(File.read(manifest_path))
+end
+
 def assert(condition, message)
   raise message unless condition
 end
@@ -36,6 +48,14 @@ def new_architecture_enabled?(options, react_native_version)
   supports_new_architecture = supports_new_architecture?(react_native_version)
   supports_new_architecture && ENV.fetch('RCT_NEW_ARCH_ENABLED',
                                          options[:fabric_enabled] || options[:turbomodule_enabled])
+end
+
+def platform_config(key, project_root, target_platform)
+  manifest = app_manifest(project_root)
+  return if manifest.nil?
+
+  config = manifest[target_platform.to_s]
+  config[key] if !config.nil? && !config.empty?
 end
 
 def resolve_module(request, start_dir = Pod::Config.instance.installation_root)
